refactor: add abstract class Table
