+++ async-race/src/app/view/pages/components/car/style.scss
